Moved rewrites into `www/js/services` - As discussed in PR #1064 - This should make merging easier down the road, and should help  clean up the `/js/` directory.
