+++ www/js/diary/LabelTab.tsx
@@ -22,7 +22,7 @@ import { SurveyOptions } from "../survey/survey";
 import { getLabelOptions } from "../survey/multilabel/confirmHelper";
 import { displayError } from "../plugin/logger";
 import { useTheme } from "react-native-paper";
-import { getPipelineRangeTs } from "../commHelper";
+import { getPipelineRangeTs } from "../services/commHelper";
 
 let labelPopulateFactory, labelsResultMap, notesResultMap, showPlaces;
 const ONE_DAY = 24 * 60 * 60; // seconds
