+++ www/js/splash/pushnotify.js
@@ -14,7 +14,7 @@
  */
 
 import angular from 'angular';
-import { updateUser } from '../commHelper';
+import { updateUser } from '../services/commHelper';
 
 angular.module('emission.splash.pushnotify', ['emission.plugin.logger',
                                               'emission.services',
